% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_wdata}
\alias{get_wdata}
\title{Retrieve weather data from WeatherAPI}
\usage{
get_wdata(
  start_date,
  end_date = start_date,
  city = NULL,
  lat = NULL,
  lng = NULL
)
}
\arguments{
\item{start_date}{Start date of the data retrieval (in "YYYY-MM-DD" format).}

\item{end_date}{End date of the data retrieval (in "YYYY-MM-DD" format). Defaults to start_date.}

\item{city}{Name of the city for which weather data is to be retrieved. Either city name or latitude and longitude should be provided.}

\item{lat}{Latitude of the location for which weather data is to be retrieved. Either city name or latitude and longitude should be provided.}

\item{lng}{Longitude of the location for which weather data is to be retrieved. Either city name or latitude and longitude should be provided.}
}
\value{
A data frame containing the retrieved weather data.
}
\description{
This function retrieves historical weather data from WeatherAPI for a given location and time period.
This API key is valid until Jun 14, 2024.
Supports start dates since 2023-01-01 and maximum time interval is 30 days (for example, from May 1 to May 31).
}
\examples{
get_wdata('2024-05-01','2024-05-31','Amsterdam')
get_wdata('2024-05-01','2024-05-31', 52.3676, 4.9041)
}
