% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wstats}
\alias{wstats}
\title{Summarizing Weather Data}
\usage{
wstats(start_date, end_date = start_date, city = NULL, lat = NULL, lng = NULL)
}
\arguments{
\item{start_date}{Start date of the data retrieval (in "YYYY-MM-DD" format).}

\item{end_date}{End date of the data retrieval (in "YYYY-MM-DD" format). Defaults to start_date.}

\item{city}{Name of the city for which weather data is to be retrieved. Either city name or latitude and longitude should be provided.}

\item{lat}{Latitude of the location for which weather data is to be retrieved. Either city name or latitude and longitude should be provided.}

\item{lng}{Longitude of the location for which weather data is to be retrieved. Either city name or latitude and longitude should be provided.}
}
\value{
a list of weather summary: maximum and minimum temperature, total precipitation, total snow, number of sunny days, and maximum wind speed
}
\description{
This function summarizes the maximum and minimum temperature, total precipitation, total snow, number of sunny days, and maximum wind speed
retrieved from WeatherAPI for a given location and time period.
}
\examples{
weather_stats <- wstats('2024-05-01','2024-05-31','Amsterdam')
weather_stats <- wstats('2024-05-01','2024-05-31', 52.3676, 4.9041)
}
\seealso{
\code{\link[=get_wdata]{get_wdata()}} for retrieving data from WeatherAPI
}
